+++ backend/app/services/catalog_service.py
@@ -141,7 +141,7 @@ def get_categorical_filters(db: Session) -> list[dict]:
 
 def get_numeric_ranges(
     db: Session,
-) -> tuple[tuple[float, float], tuple[int, int], tuple[int, int]]:
+) -> tuple[tuple[float, float], tuple[float, float], tuple[float, float]]:
     price_min, price_max = db.execute(
         select(func.min(Product.price), func.max(Product.price)).where(Product.is_active.is_(True))
     ).one()
@@ -156,8 +156,8 @@ def get_numeric_ranges(
         )
     ).one()
     price_range = (float(price_min or 0), float(price_max or 0))
-    diameter_range = (int(diameter_min or 0), int(diameter_max or 0))
-    thickness_range = (int(thickness_min or 0), int(thickness_max or 0))
+    diameter_range = (float(diameter_min or 0), float(diameter_max or 0))
+    thickness_range = (float(thickness_min or 0), float(thickness_max or 0))
     return price_range, diameter_range, thickness_range
